+++ basic/models/AircraftConfiguration.php
@@ -14,7 +14,7 @@ use Yii;
  * @property int $cargo_capacity
  * @property int $crew
  * @property int $mtow
- * @property int $bew
+ * @property int $oew
  * @property float|null $fuel_regression_a
  * @property float|null $fuel_regression_b
  * @property int|null $fuel_regression_n
@@ -41,10 +41,10 @@ class AircraftConfiguration extends \yii\db\ActiveRecord
     public function rules()
     {
         return [
-            [['aircraft_type_id', 'name', 'pax_capacity', 'cargo_capacity', 'crew', 'mtow', 'bew'], 'required'],
+            [['aircraft_type_id', 'name', 'pax_capacity', 'cargo_capacity', 'crew', 'mtow', 'oew'], 'required'],
             [['aircraft_type_id', 'pax_capacity', 'cargo_capacity'], 'integer', 'min' => 0],
-            [['crew', 'mtow', 'bew'], 'integer', 'min' => 1],
-            [['mtow'], 'compare', 'compareAttribute' => 'bew', 'operator' => '>', 'type' => 'number'],
+            [['crew', 'mtow', 'oew'], 'integer', 'min' => 1],
+            [['mtow'], 'compare', 'compareAttribute' => 'oew', 'operator' => '>', 'type' => 'number'],
             [['cargo_capacity'], 'compare', 'compareAttribute' => 'mtow', 'operator' => '<', 'type' => 'number'],
             [['name'], 'string', 'max' => 20],
             [['name'], 'trim'],
@@ -60,7 +60,7 @@ class AircraftConfiguration extends \yii\db\ActiveRecord
     public function scenarios()
     {
         $scenarios = parent::scenarios();
-        $scenarios[self::SCENARIO_ADMIN_FORM] = ['aircraft_type_id', 'name', 'pax_capacity', 'cargo_capacity', 'crew', 'mtow', 'bew'];
+        $scenarios[self::SCENARIO_ADMIN_FORM] = ['aircraft_type_id', 'name', 'pax_capacity', 'cargo_capacity', 'crew', 'mtow', 'oew'];
         return $scenarios;
     }
 
@@ -77,7 +77,7 @@ class AircraftConfiguration extends \yii\db\ActiveRecord
             'cargo_capacity' => Yii::t('app', 'Cargo Capacity (Kg)'),
             'crew' => Yii::t('app', 'Crew'),
             'mtow' => Yii::t('app', 'MTOW (Kg)'),
-            'bew' => Yii::t('app', 'BEW (Kg)'),
+            'oew' => Yii::t('app', 'OEW (Kg)'),
         ];
     }