+++ basic/tests/fixtures/data/aircraft_configuration.php
@@ -9,7 +9,7 @@ return [
         'cargo_capacity' => 4900,
         'crew' => 5,
         'mtow' => 79016,
-        'bew' => 41413,
+        'oew' => 41413,
     ],
     [
         'id' => 2,
@@ -19,7 +19,7 @@ return [
         'cargo_capacity' => 23500,
         'crew' => 3,
         'mtow' => 79016,
-        'bew' => 41413,
+        'oew' => 41413,
     ],
     [
         'id' => 3,
@@ -29,7 +29,7 @@ return [
         'cargo_capacity' => 300,
         'crew' => 1,
         'mtow' => 1111,
-        'bew' => 767,
+        'oew' => 767,
     ],
     [
         'id' => 4,
@@ -39,6 +39,6 @@ return [
         'cargo_capacity' => 100,
         'crew' => 1,
         'mtow' => 1111,
-        'bew' => 767,
+        'oew' => 767,
     ],
 ];