+++ basic/tests/unit/commands/StatisticsControllerTest.php
@@ -81,7 +81,7 @@ class StatisticsControllerTest extends BaseUnitTest
             'cargo_capacity' => 2000,
             'crew' => 5,
             'mtow' => 79016,
-            'bew' => 41413,
+            'oew' => 41413,
         ]);
         $config->save(false);